+++ src/org/bds/test/BdsTest.java
@@ -30,8 +30,8 @@ public class BdsTest {
 	public boolean verbose;
 	public boolean testCases; // Is this a bds-test? i.e. should it run as 'bds --test'?
 	public Boolean compileOk;
-	public String args[]; // Command line arguments (before program name)
-	public String argsAfter[]; // Command line arguments (after program name)
+	public String args[]; // Command line arguments (for 'bds', not for the script)
+	public String scriptArgs[]; // Command line arguments for the bds script
 	public String fileName;
 	public CompilerMessages compilerMessages;
 	public Bds bds;
@@ -49,10 +49,10 @@ public class BdsTest {
 		this(fileName, args, null, verbose, debug);
 	}
 
-	public BdsTest(String fileName, String args[], String argsAfter[], boolean verbose, boolean debug) {
+	public BdsTest(String fileName, String args[], String scriptArgs[], boolean verbose, boolean debug) {
 		this.fileName = fileName;
 		this.args = args;
-		this.argsAfter = argsAfter;
+		this.scriptArgs = scriptArgs;
 		this.verbose = verbose;
 		this.debug = debug;
 	}
@@ -88,8 +88,9 @@ public class BdsTest {
 
 		l.add(fileName); // Add file
 
-		if (argsAfter != null) {
-			for (String arg : argsAfter)
+		// Script arguments
+		if (scriptArgs != null) {
+			for (String arg : scriptArgs)
 				l.add(arg);
 		}
 
@@ -216,10 +217,16 @@ public class BdsTest {
 		checkStdout(expectedStdout, false);
 	}
 
+	/**
+	 * Check that the script's STDOUT includes 'expectedStdout' exactly one time
+	 * (or that it does NOT include it, if 'negate' is set)
+	 * @param expectedStdout
+	 * @param negate
+	 */
 	public void checkStdout(String expectedStdout, boolean negate) {
 		int count = countMatchesStdout(expectedStdout);
 
-		if (negate) Assert.assertFalse(errMsg("Error: NOT expected string '" + expectedStdout + "' in STDOUT not found"), count > 0);
+		if (negate) Assert.assertFalse(errMsg("Error: NOT expected string '" + expectedStdout + "' in STDOUT, but it was found"), count > 0);
 		else {
 			if (count <= 0) {
 				// Not found? Print differences
@@ -356,7 +363,9 @@ public class BdsTest {
 		String[] expoutLines = expout.split("\n");
 		String[] outLines = out.split("\n");
 		for (int i = 0; i < expoutLines.length; i++) {
-			if (!expoutLines[i].equals(outLines[i])) {
+			if (outLines.length <= i) {
+				System.err.println("Line " + i + "\n\tWARNING: Output does not have line number " + i + "\n\t" + expoutLines[i]);
+			} else if (!expoutLines[i].equals(outLines[i])) {
 				System.err.println("Line " + i + "\n\t" + outLines[i] + "\n\t" + expoutLines[i]);
 			}
 		}
